# Optimize cross persistent root ref updates