Update `-updateCrossPersistentRootReferencesToPersistentRoot` to use the relationship caches, so it only updates object graphs that actually have references to the target.

TestHistoryNavigationPerformance times still are not great, but at least more reasonable :-)
```
2016-01-21 02:06:06.269 Test[61866:791989] Store URL: file:///var/folders/sg/57wxmnq50t18qrgggm4h28sc0000gn/T/TestStore.sqlite
2016-01-21 02:06:06.463 Test[61866:791989] === [TestHistoryNavigationPerformance testGoToOldestAndNewestNodesInHistory] ===
2016-01-21 02:06:07.569 Test[61866:791989] In-memory snapshot is stale
2016-01-21 02:06:07.712 Test[61866:791989] Time to commit 500 new persistent roots with undo track: 1.25s
2016-01-21 02:06:14.738 Test[61866:791989] Time to make 100 commits with undo track: 7.03s
2016-01-21 02:06:15.244 Test[61866:791989] Time to go to newest node on undo track: 0.51s
2016-01-21 02:06:15.547 Test[61866:791989] Time to go to oldest node on undo track: 0.30s
2016-01-21 02:06:16.422 Test[61866:791989] Took 10145 ms
2016-01-21 02:06:16.422 Test[61866:791989] Result: 1 classes, 1 methods, 0 tests, 0 failed, 0 exceptions
```