@ericwa 2016-01-21
With an added SQL index the times are now:
```
2016-01-21 02:38:23.502 BenchmarkCoreObject[64422:820820] Store URL: file:///var/folders/sg/57wxmnq50t18qrgggm4h28sc0000gn/T/TestStore.sqlite/
2016-01-21 02:38:23.702 BenchmarkCoreObject[64422:820820] === [TestHistoryNavigationPerformance testGoToOldestAndNewestNodesInHistory] ===
2016-01-21 02:38:24.731 BenchmarkCoreObject[64422:820820] In-memory snapshot is stale
2016-01-21 02:38:24.841 BenchmarkCoreObject[64422:820820] Time to commit 500 new persistent roots with undo track: 1.14s
2016-01-21 02:38:29.848 BenchmarkCoreObject[64422:820820] Time to make 100 commits with undo track: 5.01s
2016-01-21 02:38:30.182 BenchmarkCoreObject[64422:820820] Time to go to newest node on undo track: 0.33s
2016-01-21 02:38:30.452 BenchmarkCoreObject[64422:820820] Time to go to oldest node on undo track: 0.27s
2016-01-21 02:38:31.381 BenchmarkCoreObject[64422:820820] Took 7874 ms
2016-01-21 02:38:31.381 BenchmarkCoreObject[64422:820820] Result: 1 classes, 1 methods, 0 tests, 0 failed, 0 exceptions
```